requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
email-validator
openpyxl
pypdf
python-dotenv
